osd/osd_types: pool_opts_t: int -> int64_t
Encode int32_t for pre-nautilus maps.

Fix up accessors.

+++ src/osd/osd_types.h
@@ -1029,7 +1029,7 @@ public:
     }
   };
 
-  typedef boost::variant<std::string,int,double> value_t;
+  typedef boost::variant<std::string,int64_t,double> value_t;
 
   static bool is_opt_name(const std::string& name);
   static opt_desc_t get_opt_desc(const std::string& name);
